[ADD] Command to close chats
